+++ src/commands/hardis/org/configure/monitoring.ts
@@ -50,7 +50,7 @@ export default class OrgConfigureMonitoring extends SfdxCommand {
   protected static requiresUsername = true;
 
   // Comment this out if your command does not support a hub org username
-  protected static supportsDevhubUsername = false;
+  protected static requiresDevhubUsername = false;
 
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   protected static requiresProject = false;