+++ src/commands/hardis/project/clean/xml.ts
@@ -81,7 +81,7 @@ Note: If globpattern and xpath are not sent, elements defined in property **clea
   protected static requiresUsername = false;
 
   // Comment this out if your command does not support a hub org username
-  protected static supportsDevhubUsername = false;
+  protected static requiresDevhubUsername = false;
 
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   protected static requiresProject = true;